+++ src/typography/brand/index.tsx
@@ -1,0 +1,1 @@
+export { TextBrand } from './Brand.style'